train_backend/requirements.txt
asgiref==3.6.0
Django==3.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.13.1
pytz==2023.3
sqlparse==0.4.4
